# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "支持多租户、多存储后端的软件升级管理服务",
        "title": "软件升级管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/login": {
            "post": {
                "description": "使用用户名密码登录获取JWT令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/admin/software": {
            "get": {
                "description": "分页获取软件列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "软件管理"
                ],
                "summary": "获取软件列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.ListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的软件项目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "软件管理"
                ],
                "summary": "创建软件",
                "parameters": [
                    {
                        "description": "软件信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateSoftwareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Software"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/software/{id}": {
            "get": {
                "description": "根据ID获取软件详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "软件管理"
                ],
                "summary": "获取软件详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "软件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Software"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新软件信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "软件管理"
                ],
                "summary": "更新软件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "软件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "软件信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateSoftwareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Software"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "软删除软件",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "软件管理"
                ],
                "summary": "删除软件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "软件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/admin/stats/downloads": {
            "get": {
                "description": "获取下载统计数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用量统计"
                ],
                "summary": "获取下载统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/service.StatsResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/stats/software/{id}/downloads": {
            "get": {
                "description": "获取指定软件的下载统计",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用量统计"
                ],
                "summary": "获取软件下载统计",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "软件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer",
                                            "format": "int64"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ListResponse": {
            "type": "object",
            "properties": {
                "list": {},
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.LoginResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.Software": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tenant_id": {
                    "description": "多租户预留字段",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "versions": {
                    "description": "关联",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Version"
                    }
                }
            }
        },
...
        "model.Version": {
            "type": "object",
            "properties": {
                "changelog": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_forced": {
                    "type": "boolean"
                },
                "is_incremental": {
                    "type": "boolean"
                },
                "min_version": {
                    "type": "string"
                },
                "package_hash": {
                    "type": "string"
                },
                "package_size": {
                    "type": "integer"
                },
                "package_url": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "software": {
                    "description": "关联",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Software"
                        }
                    ]
                },
                "software_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "tenant_id": {
                    "description": "多租户预留字段",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                },
                "version_code": {
                    "type": "integer"
                }
            }
        },
...
        "service.CreateSoftwareRequest": {
            "type": "object",
            "required": [
                "name",
                "platform",
                "slug"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "platform": {
                    "description": "windows/macos/linux",
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "tenant_id": {
                    "description": "多租户预留",
                    "type": "integer"
                }
            }
        },
...
        "service.StatsResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "download_count": {
                    "type": "integer"
                },
                "unique_devices": {
                    "type": "integer"
                }
            }
        },
        "service.UpdateSoftwareRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}